Creates sword packages (i.e. zip files)
closes #4

We are intentionally at this point NOT persisting these files
indefinitely (which is sort of a problem for the model to handle as it
doesn’t really expect these files to disappear all the time).

It will likely be prudent to move them to S3 if even for a short time.