Email submitter on successful submission
Handles queued vs deposited emails. The initial deposit may or may not enter a
DSpace work queue. If it is directly deposited, we can email the handle. If not,
we email a message explaining it is in the queue and an additional email
(with the handle) will be emailed when the submission is processed.

Suppressing rubocop warnings on long test classes.