Fixes bug with mets.xml
c96f8fc introduced a bug in which
DSpace was unable to parse the xml using a namespace with no
definition.

Additionally, we were unable to extract the callback URI from the XML
unless it was moved into EPDCX. I have no idea why and will leave that
as a challenge for someone in the future to solve.