Refactor from local to cloud storage part 2
Direct to S3 upload in the form via jQuery direct upload.

- Submission Create no longer creates or submits packages

Create needs to not be in the business of handling sword package creation
and submission to dspace. Moving those outside of this scope will allow
for much easier testing (and ultimately allow for a separate process to
handle that out of band with the web interface).

- We still work on a queuing jobs to create sword packages / deposit
to dspace / read response / send emails

In other words, this removed functionality that is required but will be added
back in a separate commit to reduce the size and complexity of this already
large and complex set of commits.